support decimals in json reader
